extract collection functions
